feat(import-export): Create dedicated module for data handling
-   Creates a new module at `features/import-export/data-service.js` to isolate data backup and restore logic.
-   Moves the `exportData` and `handleImport` functions from `core/main.js` into this new, dedicated module.
-   Refactors `core/main.js` to import and call the functions from the new data service.
-   Updates the event listeners for the "Export Data" and "Import Data" buttons to use the new service.
-   This change improves separation of concerns and makes the core application logic cleaner and more maintainable.
